requirements.txt
streamlit
pandas
requests
beautifulsoup4
python-dotenv
python-telegram-bot
plotly
framer-motion-py
